refactor(reduce): create a support file for reduce
Reduce can be used to implement other operators such as max and min.
Creating a support file for reduce will allow this approach.
