+++ src/operators/reduce-support.ts
@@ -1,0 +1,56 @@
+import Operator from '../Operator';
+import Subscriber from '../Subscriber';
+
+import tryCatch from '../util/tryCatch';
+import {errorObject} from '../util/errorObject';
+
+export class ReduceOperator<T, R> implements Operator<T, R> {
+
+  acc: R;
+  project: (acc: R, x: T) => R;
+
+  constructor(project: (acc: R, x: T) => R, acc?: R) {
+    this.acc = acc;
+    this.project = project;
+  }
+
+  call(subscriber: Subscriber<T>): Subscriber<T> {
+    return new ReduceSubscriber(subscriber, this.project, this.acc);
+  }
+}
+
+export class ReduceSubscriber<T, R> extends Subscriber<T> {
+
+  acc: R;
+  hasSeed: boolean;
+  hasValue: boolean = false;
+  project: (acc: R, x: T) => R;
+
+  constructor(destination: Subscriber<T>, project: (acc: R, x: T) => R, acc?: R) {
+    super(destination);
+    this.acc = acc;
+    this.project = project;
+    this.hasSeed = typeof acc !== 'undefined';
+  }
+
+  _next(x) {
+    if (this.hasValue || (this.hasValue = this.hasSeed)) {
+      const result = tryCatch(this.project).call(this, this.acc, x);
+      if (result === errorObject) {
+        this.destination.error(errorObject.e);
+      } else {
+        this.acc = result;
+      }
+    } else {
+      this.acc = x;
+      this.hasValue = true;
+    }
+  }
+
+  _complete() {
+    if (this.hasValue || this.hasSeed) {
+      this.destination.next(this.acc);
+    }
+    this.destination.complete();
+  }
+}
